# Call C level `vctrs_init()` from R level `vec_init()`