--- a/R/slice.R
+++ b/R/slice.R
@@ -650,7 +650,10 @@ vec_index <- function(x, i, ...) {
 #' vec_init(Sys.Date(), 5)
 #' vec_init(mtcars, 2)
 vec_init <- function(x, n = 1L) {
-  vec_slice(x, rep_len(NA_integer_, n))
+  n <- vec_cast(n, integer())
+  vec_assert(n, size = 1L)
+
+  .Call(vctrs_init, x, n)
 }
 
 #' Repeatedly slice a vector