--- a/src/init.c
+++ b/src/init.c
@@ -95,7 +95,7 @@ extern SEXP compact_seq(R_len_t, R_len_t, bool);
 extern SEXP init_compact_seq(int*, R_len_t, R_len_t, bool);
 
 // Defined below
-SEXP vctrs_init(SEXP);
+SEXP vctrs_init_library(SEXP);
 
 static const R_CallMethodDef CallEntries[] = {
   {"vctrs_list_get",                   (DL_FUNC) &vctrs_list_get, 2},
@@ -128,7 +128,7 @@ static const R_CallMethodDef CallEntries[] = {
   {"vctrs_compare",                    (DL_FUNC) &vctrs_compare, 3},
   {"vctrs_match",                      (DL_FUNC) &vctrs_match, 2},
   {"vctrs_typeof",                     (DL_FUNC) &vctrs_typeof, 2},
-  {"vctrs_init",                       (DL_FUNC) &vctrs_init, 1},
+  {"vctrs_init_library",               (DL_FUNC) &vctrs_init_library, 1},
   {"vctrs_is_vector",                  (DL_FUNC) &vctrs_is_vector, 1},
   {"vctrs_type2",                      (DL_FUNC) &vctrs_type2, 4},
   {"vctrs_typeof2",                    (DL_FUNC) &vctrs_typeof2, 2},
@@ -229,7 +229,7 @@ void vctrs_init_type_info(SEXP ns);
 void vctrs_init_unspecified(SEXP ns);
 void vctrs_init_utils(SEXP ns);
 
-SEXP vctrs_init(SEXP ns) {
+SEXP vctrs_init_library(SEXP ns) {
   vctrs_init_cast(ns);
   vctrs_init_data(ns);
   vctrs_init_dictionary(ns);