This is mainly useful to keep the C and R interfaces in line. It would have been easier for me to find the bug in https://github.com/r-lib/vctrs/pull/677 if the R level `vec_init()` had used a `compact_rep()`.

It does have some performance benefits when initializing atomic vectors (from not having to initialize a vector of `NA`s), less so for more complex objects.

Before:

``` r
library(vctrs)

bench::mark(vec_init(1, 1e5), iterations = 1000)
#> # A tibble: 1 x 6
#>   expression              min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>         <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_init(1, 1e+05)    270µs    364µs     2587.    1.15MB     58.2

bench::mark(vec_init(mtcars, 1e5), iterations = 200)
#> # A tibble: 1 x 6
#>   expression                   min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>              <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_init(mtcars, 1e+05)   14.5ms   19.6ms      50.3      13MB     60.2
```

<sup>Created on 2019-11-22 by the [reprex package](https://reprex.tidyverse.org) (v0.3.0.9000)</sup>

After:

``` r
library(vctrs)

bench::mark(vec_init(1, 1e5), iterations = 1000)
#> # A tibble: 1 x 6
#>   expression              min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>         <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_init(1, 1e+05)   66.2µs    157µs     4523.     861KB     59.6

bench::mark(vec_init(mtcars, 1e5), iterations = 200)
#> # A tibble: 1 x 6
#>   expression                   min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>              <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_init(mtcars, 1e+05)   14.4ms     19ms      52.7    12.6MB     75.4
```

<sup>Created on 2019-11-22 by the [reprex package](https://reprex.tidyverse.org) (v0.3.0.9000)</sup>